update for all the routes
